refactor: merge `messages/receive.py` module
