--- src/piper_kit/messages/receive/receive.py
@@ -1,8 +1,0 @@
-"""Base class for CAN messages received from the PiPER arm."""
-
-
-class ReceiveMessage:
-    """Base class for CAN messages received from the PiPER robotic arm."""
-
-
-__all__ = ["ReceiveMessage"]
